+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 * Unreleased
+    * Validate that the zoneId and linkId cannot be 0x00, because 0x00
+      is used as an error return code in certain parts of the AceTime C++ code.
 * v0.1 (2021-10-06)
     * Extract `zone_processor.py` and `acetz.py` to new
       [AceTimePython](https://github.com/bxparks/AceTimePython) library.